Wing morphology of mutant flies

Fold changes of different measures of wing morphology in heterozygous (Tps1MIC/+) and homozygous (Tps1MIC) Tps1 mutant flies. Data are expressed as percentage change relative to the mean of the heterozygous mutants.

wings

Format

An object of class tbl_df (inherits from tbl, data.frame) with 120 rows and 4 columns.

Details

40 flies were measured in total, with 3 measurements taken per fly.

sex

factor. Male or female.

genotype

factor. Heterozygous (Tps1MIC/+) or homozygous (Tps1MIC) mutant

measure

factor. Type of wing measurement: wing size, cell size, or cell number

percent.change

double. Value measured.

References

Matsushita, R, Nishimura, T. Trehalose metabolism confers developmental robustness and stability in Drosophila by regulating glucose homeostasis. Commun Biol 3, 170 (2020). doi:10.1038/s42003-020-0889-1