Grob for offset axes with minor ticks

Grob for offset axes with minor ticks

draw_prism_offset_minor(
  break_positions,
  break_labels,
  breaks_major,
  axis_position,
  theme,
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1
)

Arguments

break_positions
position of ticks

break_labels
labels at ticks

breaks_major
logical vector indicating major ticks versus minor ticks

axis_position
position of axis (top, bottom, left or right)

theme
A complete theme object

check.overlap
silently remove overlapping labels, (recursively) prioritizing the first, last, and middle labels.

angle
Compared to setting the angle in theme / element_text, this also uses some heuristics to automatically pick the hjust and vjust that you probably want.

n.dodge
The number of rows (for vertical axes) or columns (for horizontal axes) that should be used to render the labels. This is useful for displaying labels that would otherwise overlap.